--- examples/react-app/.env.example
@@ -1,4 +1,0 @@
-PORT=5173
-VITE_FUEL_PROVIDER_URL="https://testnet.fuel.network/v1/graphql"
-VITE_WALLET_SECRET= # leave it empty to skip seed wallet
-VITE_MASTER_WALLET_MNEMONIC=
