--- examples/react-app/playwright/e2e/utils/index.ts
@@ -1,1 +1,0 @@
-export * from './connect';
